requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database and search
elasticsearch==8.11.0
elasticsearch-dsl==8.11.0
redis==5.0.1

# Machine learning and AI
sentence-transformers==2.2.2
openai==1.3.5

# File processing
python-docx==1.1.0
PyPDF2==3.0.1
beautifulsoup4==4.12.2
aiofiles==23.2.1

# Configuration and utilities
python-dotenv==1.0.0
pydantic-settings==2.11.0
pydantic==2.7.0
pathlib2==2.3.7  # For older Python versions, if needed
apscheduler==3.11.0
httpx==0.24.1  # 추가: httpx 버전 고정 (OpenAI 호환)
python-jose[cryptography]==3.3.0
